refactor: 설정 화면을 별도의 패키지로 옮김
